requirements.txt
psycopg==3.2.12
Flask==3.1.2
MarkupSafe==3.0.3
python-dotenv==1.2.1